# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Self-Service Kiosk API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Categoria {
  id: number;
  nome: string;
  icone?: string | null;
  imagemUrl?: string | null;
  ordem: number;
}

export interface CriarCategoriaInput {
  nome: string;
  icone?: string | null;
  imagemUrl?: string | null;
  ordem: number;
}

export type ExtraTipo = (typeof ExtraTipo)[keyof typeof ExtraTipo];

export const ExtraTipo = {
  adicional: "adicional",
  remocao: "remocao",
  tamanho: "tamanho",
} as const;

export type ExtraModoSelecao =
  (typeof ExtraModoSelecao)[keyof typeof ExtraModoSelecao];

export const ExtraModoSelecao = {
  multipla: "multipla",
  unica: "unica",
  sim_nao: "sim_nao",
} as const;

export interface Extra {
  id: number;
  nome: string;
  preco: number;
  tipo: ExtraTipo;
  grupoTitulo: string;
  modoSelecao: ExtraModoSelecao;
  maxSelecoes: number;
  obrigatorio: boolean;
  ordemGrupo: number;
  ordemItem: number;
  /** URL da foto da opção no totem (/api/uploads/...) */
  imagem?: string | null;
}

export interface Produto {
  id: number;
  nome: string;
  descricao?: string | null;
  preco: number;
  categoriaId: number;
  imagem?: string | null;
  ativo: boolean;
  extras: Extra[];
  categoria?: Categoria | null;
}

export type CriarExtraInputTipo =
  (typeof CriarExtraInputTipo)[keyof typeof CriarExtraInputTipo];

export const CriarExtraInputTipo = {
  adicional: "adicional",
  remocao: "remocao",
  tamanho: "tamanho",
} as const;

export type CriarExtraInputModoSelecao =
  (typeof CriarExtraInputModoSelecao)[keyof typeof CriarExtraInputModoSelecao];

export const CriarExtraInputModoSelecao = {
  multipla: "multipla",
  unica: "unica",
  sim_nao: "sim_nao",
} as const;

export interface CriarExtraInput {
  nome: string;
  preco: number;
  tipo: CriarExtraInputTipo;
  grupoTitulo?: string;
  modoSelecao?: CriarExtraInputModoSelecao;
  maxSelecoes?: number;
  obrigatorio?: boolean;
  ordemGrupo?: number;
  ordemItem?: number;
  imagem?: string | null;
}

export interface CriarProdutoInput {
  nome: string;
  descricao?: string | null;
  preco: number;
  categoriaId: number;
  imagem?: string | null;
  ativo: boolean;
  extras?: CriarExtraInput[];
}

export interface ItemPedido {
  id: number;
  pedidoId: number;
  produtoId: number;
  quantidade: number;
  preco: number;
  observacoes?: string | null;
  extrasIds: number[];
  produto?: Produto | null;
}

export interface CriarItemPedidoInput {
  produtoId: number;
  quantidade: number;
  observacoes?: string | null;
  extrasIds: number[];
}

export type PedidoStatus = (typeof PedidoStatus)[keyof typeof PedidoStatus];

export const PedidoStatus = {
  novo: "novo",
  em_preparo: "em_preparo",
  pronto: "pronto",
  entregue: "entregue",
  cancelado: "cancelado",
} as const;

export type PedidoTipoPedido =
  (typeof PedidoTipoPedido)[keyof typeof PedidoTipoPedido];

export const PedidoTipoPedido = {
  comer_aqui: "comer_aqui",
  para_viagem: "para_viagem",
} as const;

export type PedidoFormaPagamento =
  | (typeof PedidoFormaPagamento)[keyof typeof PedidoFormaPagamento]
  | null;

export const PedidoFormaPagamento = {
  credito: "credito",
  debito: "debito",
  pix: "pix",
  nfc: "nfc",
} as const;

export interface Pedido {
  id: number;
  numero: string;
  status: PedidoStatus;
  total: number;
  tipoPedido: PedidoTipoPedido;
  formaPagamento?: PedidoFormaPagamento;
  criadoEm: string;
  itens: ItemPedido[];
}

export type CriarPedidoInputTipoPedido =
  (typeof CriarPedidoInputTipoPedido)[keyof typeof CriarPedidoInputTipoPedido];

export const CriarPedidoInputTipoPedido = {
  comer_aqui: "comer_aqui",
  para_viagem: "para_viagem",
} as const;

export type CriarPedidoInputFormaPagamento =
  (typeof CriarPedidoInputFormaPagamento)[keyof typeof CriarPedidoInputFormaPagamento];

export const CriarPedidoInputFormaPagamento = {
  credito: "credito",
  debito: "debito",
  pix: "pix",
  nfc: "nfc",
} as const;

export interface CriarPedidoInput {
  tipoPedido: CriarPedidoInputTipoPedido;
  formaPagamento: CriarPedidoInputFormaPagamento;
  itens: CriarItemPedidoInput[];
}

export type AtualizarStatusInputStatus =
  (typeof AtualizarStatusInputStatus)[keyof typeof AtualizarStatusInputStatus];

export const AtualizarStatusInputStatus = {
  novo: "novo",
  em_preparo: "em_preparo",
  pronto: "pronto",
  entregue: "entregue",
  cancelado: "cancelado",
} as const;

export interface AtualizarStatusInput {
  status: AtualizarStatusInputStatus;
}

export interface ProdutoVendido {
  produto: string;
  quantidade: number;
  total: number;
}

export interface VendaPorHora {
  hora: string;
  total: number;
  quantidade: number;
}

export interface RelatorioVendas {
  totalVendas: number;
  totalPedidos: number;
  ticketMedio: number;
  produtosMaisVendidos: ProdutoVendido[];
  vendasPorHora: VendaPorHora[];
}

export type ImpressoraTipoConexao =
  (typeof ImpressoraTipoConexao)[keyof typeof ImpressoraTipoConexao];

export const ImpressoraTipoConexao = {
  rede: "rede",
  usb: "usb",
} as const;

export type ImpressoraLarguraPapel =
  (typeof ImpressoraLarguraPapel)[keyof typeof ImpressoraLarguraPapel];

export const ImpressoraLarguraPapel = {
  "80mm": "80mm",
  "58mm": "58mm",
} as const;

export interface Impressora {
  id: number;
  nome: string;
  tipoConexao: ImpressoraTipoConexao;
  endereco: string;
  larguraPapel: ImpressoraLarguraPapel;
  margemEsquerda: number;
  margemDireita: number;
  ativa: boolean;
  criadoEm: string;
  atualizadoEm: string;
}

export type CriarImpressoraInputTipoConexao =
  (typeof CriarImpressoraInputTipoConexao)[keyof typeof CriarImpressoraInputTipoConexao];

export const CriarImpressoraInputTipoConexao = {
  rede: "rede",
  usb: "usb",
} as const;

export type CriarImpressoraInputLarguraPapel =
  (typeof CriarImpressoraInputLarguraPapel)[keyof typeof CriarImpressoraInputLarguraPapel];

export const CriarImpressoraInputLarguraPapel = {
  "80mm": "80mm",
  "58mm": "58mm",
} as const;

export interface CriarImpressoraInput {
  nome: string;
  tipoConexao?: CriarImpressoraInputTipoConexao;
  endereco: string;
  larguraPapel?: CriarImpressoraInputLarguraPapel;
  margemEsquerda?: number;
  margemDireita?: number;
  ativa?: boolean;
}

export type ListarProdutosParams = {
  categoria?: number;
  ativo?: boolean;
};

export type ListarPedidosParams = {
  status?: ListarPedidosStatus;
  data?: string;
};

export type ListarPedidosStatus =
  (typeof ListarPedidosStatus)[keyof typeof ListarPedidosStatus];

export const ListarPedidosStatus = {
  novo: "novo",
  em_preparo: "em_preparo",
  pronto: "pronto",
  entregue: "entregue",
  cancelado: "cancelado",
} as const;

export type EmitirNfceBodyPayload = { [key: string]: unknown };

export type EmitirNfceBody = {
  payload?: EmitirNfceBodyPayload;
};

export type RelatorioVendasParams = {
  data?: string;
};