# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Restaurant Self-Service Kiosk API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fiscal/nfce/emitir/{pedidoId}:
    post:
      operationId: emitirNfce
      tags: [fiscal]
      summary: Emitir NFC-e para um pedido
      parameters:
        - name: pedidoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
      responses:
        "202":
          description: Emissão aceita ou em processamento
        "500":
          description: Internal server error
...
  /impressoras:
    get:
      operationId: listarImpressoras
      tags: [impressoras]
      summary: List all printers
      responses:
        "200":
          description: List of printers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Impressora"
    post:
      operationId: criarImpressora
      tags: [impressoras]
      summary: Create a printer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarImpressoraInput"
      responses:
        "201":
          description: Printer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Impressora"
  /impressoras/{id}:
    put:
      operationId: atualizarImpressora
      tags: [impressoras]
      summary: Update a printer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarImpressoraInput"
      responses:
        "200":
          description: Printer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Impressora"
    delete:
      operationId: excluirImpressora
      tags: [impressoras]
      summary: Delete a printer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Printer deleted
  /impressoras/{id}/test:
    post:
      operationId: testarImpressora
      tags: [impressoras]
      summary: Print test page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test successful
components:
  schemas:
...
    Impressora:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        tipoConexao:
          type: string
          enum: [rede, usb]
        endereco:
          type: string
        larguraPapel:
          type: string
          enum: [80mm, 58mm]
        margemEsquerda:
          type: integer
        margemDireita:
          type: integer
        ativa:
          type: boolean
        criadoEm:
          type: string
        atualizadoEm:
          type: string
      required:
        - id
        - nome
        - tipoConexao
        - endereco
        - larguraPapel
        - margemEsquerda
        - margemDireita
        - ativa
        - criadoEm
        - atualizadoEm
    CriarImpressoraInput:
      type: object
      properties:
        nome:
          type: string
        tipoConexao:
          type: string
          enum: [rede, usb]
        endereco:
          type: string
        larguraPapel:
          type: string
          enum: [80mm, 58mm]
        margemEsquerda:
          type: integer
        margemDireita:
          type: integer
        ativa:
          type: boolean
      required:
        - nome
        - endereco