# lib/api-zod/src/generated/types/criarImpressoraInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Self-Service Kiosk API
 * OpenAPI spec version: 0.1.0
 */
import type { CriarImpressoraInputLarguraPapel } from "./criarImpressoraInputLarguraPapel";
import type { CriarImpressoraInputTipoConexao } from "./criarImpressoraInputTipoConexao";

export interface CriarImpressoraInput {
  nome: string;
  tipoConexao?: CriarImpressoraInputTipoConexao;
  endereco: string;
  larguraPapel?: CriarImpressoraInputLarguraPapel;
  margemEsquerda?: number;
  margemDireita?: number;
  ativa?: boolean;
}