# lib/api-zod/src/generated/types/impressora.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Self-Service Kiosk API
 * OpenAPI spec version: 0.1.0
 */
import type { ImpressoraLarguraPapel } from "./impressoraLarguraPapel";
import type { ImpressoraTipoConexao } from "./impressoraTipoConexao";

export interface Impressora {
  id: number;
  nome: string;
  tipoConexao: ImpressoraTipoConexao;
  endereco: string;
  larguraPapel: ImpressoraLarguraPapel;
  margemEsquerda: number;
  margemDireita: number;
  ativa: boolean;
  criadoEm: string;
  atualizadoEm: string;
}